# openapi.yaml
openapi: 3.0.3
info:
  title: Blockvisor API
  description: |
    Blockvisor API is a distributed infrastructure management system that orchestrates software deployment and monitoring across physical and virtual machines. Originally designed for blockchain node management, it has evolved into a general-purpose platform for running and managing any software.
    
    ## Core Concepts
    - **Hosts**: Physical or virtual machines registered with the system via the `bvup` tool
    - **Nodes**: Running software instances with real-time state and metrics monitoring
    - **Organizations**: Groups of users sharing resources and permissions
    - **Users**: Individual accounts with role-based access control (RBAC)
    - **Commands**: Instructions sent to hosts for software lifecycle management
    
    ## Authentication
    The API uses JWT tokens for authentication. Include the token in the `Authorization` header as `Bearer <token>`.
    
    ## Error Handling
    The API returns structured error responses with detailed parameter validation information:
    ```json
    {
      "error": "Invalid query parameters",
      "details": [
        {
          "parameter": "org_id",
          "error": "Invalid UUID format: 'not-a-uuid'",
          "expected": "Valid UUID string (e.g., '550e8400-e29b-41d4-a716-446655440000')"
        }
      ]
    }
    ```
  version: 0.4.0
  contact:
    name: Blockvisor API Support
  license:
    name: MIT
servers:
  - url: https://api.blockvisor.com
...
security:
  - BearerAuth: []
paths:
...
  /v1/org:
    get:
      tags:
        - Organizations
      summary: List organizations
      description: Get a list of organizations with optional filtering
      parameters:
        - name: member_id
          in: query
          description: Filter by member ID - only return orgs that this user is a member of
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: personal
          in: query
          description: If true, only personal orgs are returned, otherwise none are
          schema:
            type: boolean
            example: false
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
            example: 50
      responses:
        '200':
          description: Organizations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
                  total:
                    type: integer
                    description: Total number of organizations
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

    post:
      tags:
        - Organizations
      summary: Create organization
      description: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "My Organization"
                  minLength: 1
                  maxLength: 255
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    $ref: '#/components/schemas/Organization'
  /v1/org/{id}:
    get:
      tags:
        - Organizations
      summary: Get organization
      description: Get details of a specific organization
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Organization retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    $ref: '#/components/schemas/Organization'
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Organizations
      summary: Update organization
      description: Update an existing organization
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Updated Organization Name"
                  minLength: 1
                  maxLength: 255
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    $ref: '#/components/schemas/Organization'

    delete:
      tags:
        - Organizations
      summary: Delete organization
      description: Delete an organization
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Organization deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Organization deleted successfully"

  # Node endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "An error occurred"
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid query parameters"
        details:
          type: array
          items:
            type: object
            properties:
              parameter:
                type: string
                example: "org_id"
              error:
                type: string
                example: "Invalid UUID format: 'not-a-uuid'"
              expected:
                type: string
                example: "Valid UUID string (e.g., '550e8400-e29b-41d4-a716-446655440000')"

    # Core entity schemas
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "My Organization"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        personal:
          type: boolean
          example: false
          description: Whether this is a personal organization
...